refactor(kernel): split the projection read contracts out of the rebuildable projection
